src/gcp cloud functions/requirements_1.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
cloudstorage